refactor(all): big refactoring
