+++ opis-manager/src/app/config.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
+import { take, map, tap } from 'rxjs/operators';
 
 export interface Config {
   apiUrl: string;
@@ -12,9 +13,27 @@ export interface Config {
 })
 export class ConfigService {
 
-  constructor(private http: HttpClient) { }
+  config: Config = {
+    apiUrl: '',
+    years: [],
+  };
 
-  getConfig(): Observable<object> {
-    return this.http.get('assets/config.json');
+  constructor(private readonly http: HttpClient) {
+    this.updateConfig();
   }
+
+  getConfig(): Observable<Config> {
+    return this.http.get<Config>('assets/config.json');
+  }
+
+  updateConfig(): Observable<Config> {
+    return this.getConfig().pipe(
+      take(1),
+      tap((c: Config) => {
+        this.config.apiUrl = c.apiUrl;
+        this.config.years = c.years;
+      })
+    );
+  }
+
 }
