+++ opis-manager/src/app/graph.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { GraphService } from './graph.service';
+
+describe('GraphService', () => {
+  let service: GraphService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(GraphService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
