+++ opis-manager/src/app/teaching/teaching.component.scss
@@ -1,0 +1,34 @@
+$color_main: #00897b;
+
+.nav-item-right {
+  @extend .nav-item !optional;
+  margin-left: auto;
+}
+
+
+/* Slider */
+.slider-years {
+  width: 500px;
+}
+.custom-slider {
+  margin: 0 auto;
+  display: table;
+}
+::ng-deep {
+  .custom-slider .ng5-slider .ng5-slider-bar {
+    background: #ffe4d1;
+    height: 2px;
+  }
+
+  .custom-slider .ng5-slider .ng5-slider-pointer {
+    background-color: $color_main;
+  }
+
+  .custom-slider .ng5-slider .ng5-slider-bubble {
+    bottom: 14px;
+  }
+
+  .custom-slider .ng5-slider .ng5-slider-tick.ng5-slider-selected {
+    background: $color_main;
+  }
+}
