+++ opis-manager/src/app/value-details/value-details.component.scss
@@ -1,14 +1,3 @@
-$color_dark: #00695c;
-$color_light: #fff;
-$color_main: #00897b;
-$font_family_1: "Helvetica Neue";
-
-.btn-opis {
-  color: $color_light;
-  background-color: $color_main;
-  border-color: $color_dark;
-}
-
 .v-info {
   margin: 0 auto;
   width: 900px;
