+++ opis-manager/src/styles.scss
@@ -1,22 +1,15 @@
 @import '~katex/dist/katex.css';
 
-$color_1: #167efb;
-$color_dark: #80cbc4;
+$color_dark: #00695c;
 $color_light: #fff;
 $color_main: #00897b;
 $font_family_1: "Helvetica Neue";
 
 /*
-
-#00897b  Main
-#80cbc4 Light
-#00695c Dark
+  #00897b  Main
+  #80cbc4 Light
+  #00695c Dark
 */
-
-
-body {
-	background-color: #80cbc4;
-}
 
 .container-fluid {
 	padding-right: 0;
@@ -86,4 +79,28 @@ router-outlet ~ * {
 	position: absolute;
 	height: 100%;
 	width: 100%;
+}
+
+.btn-opis {
+  color: $color_light;
+  background-color: $color_main;
+  border-color: $color_dark;
+}
+.btn-opis:hover, .btn-opis option {
+  color: $color_light;
+}
+
+.selected-tab {
+  background-color: $color_light !important;
+  color: $color_dark !important;
+}
+
+.list-option-container {
+  margin-left: 325px;
+  width: 850px;
+}
+
+.addons-bar {
+  margin-top: 5px;
+  margin-left: 20px;
 }
